Add HttpDataManagerInitializationException (Runtime), throw that on MalformedURLException
This is in contrast to marking methods as `throws MalformedURLException`.

Rationale: once we aren't in HttpDataManager anymore, the MalformedURLException becomes less useful, since it's only relevant to the HttpDataManager. Also, the URL to elasticsearch is a hard-coded string, so if it throws MalformedURL, something is very wrong and we should abort.

HttpDataManagerInitializationException is a RuntimeException, so it doesn't need to be caught, it will just halt the application immediately.